Move form-builder header to globals
