+++ components/globals/index.ts
@@ -7,3 +7,4 @@ export { default as SkipLink } from "./SkipLink";
 export { default as LanguageToggle } from "./LanguageToggle";
 export { default as Brand } from "./Brand";
 export { Alert, ErrorStatus } from "./Alert";
+export { Header } from "./Header";
